+++ crates/mesh-llm-host-runtime/Cargo.toml
@@ -10,7 +10,7 @@ gpu-bench-intel = ["mesh-llm-system/gpu-bench-intel"]
 
 [dependencies]
 ansi-to-tui = "8"
-moa-gateway = { path = "../moa-gateway" }
+mesh-mixture-of-agents = { path = "../mesh-mixture-of-agents" }
 mesh-llm-plugin = { path = "../mesh-llm-plugin" }
 mesh-llm-identity = { path = "../mesh-llm-identity" }
 mesh-llm-protocol = { path = "../mesh-llm-protocol" }
