+++ crates/mesh-mixture-of-agents/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
-name = "moa-gateway"
+name = "mesh-mixture-of-agents"
 version = "0.1.0"
 edition = "2021"
-description = "Mixture-of-Agents gateway — fan out to heterogeneous LLMs, arbitrate, return one response"
+description = "Mixture-of-Agents — fan out to heterogeneous LLMs, arbitrate, return one response"
 
 [dependencies]
+async-trait = "0.1"
 reqwest = { version = "0.12", features = ["json", "stream"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -12,15 +13,3 @@ tokio = { version = "1", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 regex = "1"
-
-[[bin]]
-name = "moa-test"
-path = "src/bin/moa_test.rs"
-
-[[bin]]
-name = "moa-mesh"
-path = "src/bin/moa_mesh.rs"
-
-[[bin]]
-name = "moa-agent"
-path = "src/bin/moa_agent.rs"
