+++ crates/mesh-mixture-of-agents/src/normalize.rs
